requirements.txt
django==4.0.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
django-livesync==0.5
mongoengine==0.24.2
django-environ==0.9.0
pymongo==4.2.0
django-rest-framework-mongoengine==3.4.1